refactor(composer): keep wheel control local
