+++ src/renderer/components/composer/variants/shared/__tests__/ComposerSpeedControl.ui.test.tsx
@@ -1,6 +1,6 @@
 import type { ThinkingOption } from '@renderer/types/reasoning'
 import { type Model, MODEL_CAPABILITY } from '@shared/data/types/model'
-import { fireEvent, render, screen, waitFor } from '@testing-library/react'
+import { createEvent, fireEvent, render, screen, waitFor } from '@testing-library/react'
 import userEvent from '@testing-library/user-event'
 import { type ButtonHTMLAttributes, type MouseEvent, type ReactNode, useState } from 'react'
 import { describe, expect, it, vi } from 'vitest'
@@ -138,7 +138,6 @@ describe('ComposerSpeedControl UI', () => {
     const slider = screen.getByTestId('reasoning-slider')
     expect(slider).toHaveAttribute('data-max', '5')
     expect(slider).toHaveAttribute('data-value', '3')
-    expect(container.querySelector('[data-slot="wheel-step-control"]')).toBeInTheDocument()
     expect(container.querySelectorAll('[data-slot="composer-effort-step"]')).toHaveLength(5)
     expect(container.querySelector('[data-slot="composer-effort-step"][data-index="3"]')).not.toBeInTheDocument()
     expect(screen.queryByTestId('reasoning-menu')).not.toBeInTheDocument()
@@ -268,6 +267,46 @@ describe('ComposerSpeedControl UI', () => {
     expect(screen.getByTestId('composer-effort-slider-label')).toHaveTextContent(
       'assistants.settings.reasoning_effort.default'
     )
+  })
+
+  it('changes one effort level per wheel step without scrolling the page', () => {
+    const outerWheel = vi.fn()
+    render(
+      <div onWheel={outerWheel}>
+        <ControlledSpeedControl model={codexModel} initialEffort="high" />
+      </div>
+    )
+
+    const slider = screen.getByRole('slider', { name: 'agent.speed.effort' })
+    const firstWheel = createEvent.wheel(slider, { deltaY: -15, cancelable: true })
+    fireEvent(slider, firstWheel)
+    expect(firstWheel.defaultPrevented).toBe(true)
+    fireEvent.wheel(slider, { deltaY: -15 })
+    expect(slider).toHaveAttribute('data-value', '3')
+
+    fireEvent.wheel(slider, { deltaY: -15 })
+    expect(slider).toHaveAttribute('data-value', '4')
+
+    fireEvent.wheel(slider, { deltaY: -100 })
+    expect(slider).toHaveAttribute('data-value', '5')
+
+    fireEvent.wheel(slider, { deltaY: -100 })
+    expect(slider).toHaveAttribute('data-value', '5')
+
+    fireEvent.wheel(slider, { deltaY: 100 })
+    expect(slider).toHaveAttribute('data-value', '4')
+    expect(outerWheel).toHaveBeenCalledTimes(1)
+  })
+
+  it('treats line and page wheel events as one effort step each', () => {
+    render(<ControlledSpeedControl model={codexModel} initialEffort="high" />)
+
+    const slider = screen.getByRole('slider', { name: 'agent.speed.effort' })
+    fireEvent.wheel(slider, { deltaMode: WheelEvent.DOM_DELTA_LINE, deltaY: -1 })
+    expect(slider).toHaveAttribute('data-value', '4')
+
+    fireEvent.wheel(slider, { deltaMode: WheelEvent.DOM_DELTA_PAGE, deltaY: -1 })
+    expect(slider).toHaveAttribute('data-value', '5')
   })
 
   it('toggles Fast only for a capable provider-model pair', () => {
